requirements.txt
PyYAML
urllib3
requests
tqdm